+++ src/board/system76/common/smbus.c
@@ -22,10 +22,10 @@ void smbus_init(void) {
     i2c_reset(&I2C_SMBUS, true);
 }
 
-int smbus_read(uint8_t address, uint8_t command, uint16_t * data) {
+int16_t smbus_read(uint8_t address, uint8_t command, uint16_t * data) {
     return i2c_get(&I2C_SMBUS, address, command, (uint8_t *)data, 2);
 }
 
-int smbus_write(uint8_t address, uint8_t command, uint16_t data) {
+int16_t smbus_write(uint8_t address, uint8_t command, uint16_t data) {
     return i2c_set(&I2C_SMBUS, address, command, (uint8_t *)&data, 2);
 }
